modified:   costmap/include/costmap/costmap.h 		Modifications to Initialize costmap 	modified:   costmap/include/costmap/costmap_ros.h 		Added variables 	new file:   costmap/include/costmap/map_cell.h 		Initial commit 	modified:   costmap/plugins/inflation_layer.cpp 		changed node name 	modified:   costmap/plugins/map_layer.cpp 		changed node name 	modified:   costmap/plugins/obstacle_layer.cpp 		changed node name 	modified:   costmap/src/costmap.cpp 		Modifications to Initialize costmap 	modified:   costmap/src/costmap_ros.cpp 		Map update loop added (multithreaded) 	new file:   costmap/src/map_cell.cpp 		Initial commit

+++ costmap/include/costmap/costmap.h
@@ -11,7 +11,7 @@ namespace costmap
 class Costmap
 {
  public:
-  Costmap(std::string global_frame, std::string base_frame, double size_x, double size_y, double resolution);
+  Costmap(std::string global_frame, std::string base_frame, unsigned int size_x, unsigned int size_y, double resolution);
   virtual ~Costmap();
 
  protected:
